Combined all my separate functions
A lot of my functions were doing nearly the same thing, so I wrote a
coupe if/else statements to pare down the code to just one long
function and a couple shorter button functions. Also, downloaded the
background I was using so that it’s still visible even if I’m working
offline. Also, starting to work on a node that tells the user whose
feed they’re viewing.